♻️ refactor: move plugin modules into core/plugin_discovery subpackage
- Move plugin_categories.py, plugin_registry.py, plugin_manager.py into app/core/plugin_discovery/
- Update all imports in pipeline_engine.py and test files
- Fix PLUGINS_DIR path for new directory depth
