+++ src/app/core/plugin_discovery/plugin_categories.py
