+++ src/app/core/plugin_discovery/plugin_manager.py
