+++ src/tests/app/integration/test_plugin_registry.py
@@ -8,8 +8,8 @@ Tests are written to be resilient to plugin changes — they discover
 plugins dynamically rather than hard-coding specific names or params.
 """
 
-from app.core.plugin_categories import PLUGIN_CATEGORIES
-from app.core.plugin_registry import (
+from app.core.plugin_discovery.plugin_categories import PLUGIN_CATEGORIES
+from app.core.plugin_discovery.plugin_registry import (
     PLUGINS_DIR,
     _find_plugin_modules,
     get_plugin_registry,
